pyproject.toml
[project]
name = "the-watcher"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "python-dotenv>=1.0.0",
    "google>=3.0.0",
    "google-genai>=1.73.1",
]
